# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RFQ Management System API for Cortoba Supplies
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Employee login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "401":
          description: Unauthorized

  # EMPLOYEES
  /employees:
    get:
      operationId: listEmployees
      tags: [auth]
      summary: List all employees
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [auth]
      summary: Create employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    patch:
      operationId: updateEmployee
      tags: [auth]
      summary: Update employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"

  # SUPPLIERS
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [employee, token]
      properties:
        employee:
          $ref: "#/components/schemas/Employee"
        token:
          type: string
    Employee:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [superadmin, admin, manager, purchasing, data_entry]
        phone:
          type: ["string", "null"]
        isActive:
          type: boolean
        permissions:
          type: ["object", "null"]
          description: "Per-employee permission map { permissionKey: true }. null = role default."
          additionalProperties:
            type: boolean
        createdAt:
          type: string
    EmployeeInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [superadmin, admin, manager, purchasing, data_entry]
        phone:
          type: string
        permissions:
          type: ["object", "null"]
          description: "Per-employee permission map { permissionKey: true }. null = role default."
          additionalProperties:
            type: boolean
    EmployeeUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [superadmin, admin, manager, purchasing, data_entry]
        phone:
          type: string
        isActive:
          type: boolean
        password:
          type: string
        permissions:
          type: ["object", "null"]
          description: "Per-employee permission map { permissionKey: true }. null = role default."
          additionalProperties:
            type: boolean

    # CATEGORIES
...